requirements.txt
streamlit==1.37.1
streamlit-authenticator==0.4.2
PyYAML==6.0.1
bcrypt==3.2.0

